my_services/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery[redis]==5.2.7
flower==1.2.0
